biometric_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
Pillow==10.1.0
